requirements.txt
aiohttp==3.7.4.post0
anyio==3.3.4
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
discord==1.7.3
discord.py==1.7.3
fastapi==0.70.0
greenlet==1.1.2
h11==0.12.0
idna==3.3
multidict==6.0.2
mysqlclient==2.1.0
numpy==1.22.2
pandas==1.4.1
pydantic==1.8.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
PyYAML==6.0
requests==2.27.1
riotwatcher==3.2.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.27
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.8
uvicorn==0.15.0
watchgod==0.7
websockets==10.0
yarl==1.7.2
